Restructure of pages
