+++ backend/predictions/3p_predictions.csv
@@ -293,15 +293,28 @@
 291,4/3/2022,Phoenix Suns,Oklahoma City Thunder,26,31,28,28,26,24.5,27.8,105,-145
 292,4/3/2022,Portland Trail Blazers,San Antonio Spurs,25,26,25,27,24,24.5,25.4,-105,-130
 293,4/3/2022,Washington Wizards,Boston Celtics,24,26,25,25,25,24.5,25,100,-140
-294,4/5/2022 0:00,Atlanta Hawks,Toronto Raptors,25,24,25,24,26,24.5,24.8,-125,-110
-295,4/5/2022 0:00,Charlotte Hornets,Miami Heat,26,28,27,27,26,26.5,26.8,-140,105
-296,4/5/2022 0:00,Cleveland Cavaliers,Orlando Magic,24,25,25,24,24,24.5,24.4,110,-150
-297,4/5/2022 0:00,Houston Rockets,Brooklyn Nets,26,25,26,26,26,24.5,25.8,-150,110
-298,4/5/2022 0:00,Los Angeles Lakers,Phoenix Suns,22,24,22,23,23,24.5,22.8,110,-150
-299,4/5/2022 0:00,Memphis Grizzlies,Utah Jazz,25,28,26,25,26,26.5,26,-105,-130
-300,4/5/2022 0:00,Milwaukee Bucks,Chicago Bulls,23,24,22,23,22,24.5,22.8,-125,-105
-301,4/5/2022 0:00,New Orleans Pelicans,Sacramento Kings,22,24,22,23,21,21.5,22.4,-135,-105
-302,4/5/2022 0:00,Philadelphia 76ers,Indiana Pacers,25,25,25,25,26,24.5,25.2,110,-150
-303,4/5/2022 0:00,Portland Trail Blazers,Oklahoma City Thunder,27,30,28,28,25,24.5,27.6,-155,115
-304,4/5/2022 0:00,San Antonio Spurs,Denver Nuggets,25,27,25,25,25,24.5,25.4,-105,-135
-305,4/5/2022 0:00,Washington Wizards,Minnesota Timberwolves,25,27,26,26,26,24.5,26,-140,105
+294,4/5/2022,Atlanta Hawks,Toronto Raptors,25,24,25,24,26,24.5,24.8,-125,-110
+295,4/5/2022,Charlotte Hornets,Miami Heat,26,28,27,27,26,26.5,26.8,-140,105
+296,4/5/2022,Cleveland Cavaliers,Orlando Magic,24,25,25,24,24,24.5,24.4,110,-150
+297,4/5/2022,Houston Rockets,Brooklyn Nets,26,25,26,26,26,24.5,25.8,-150,110
+298,4/5/2022,Los Angeles Lakers,Phoenix Suns,22,24,22,23,23,24.5,22.8,110,-150
+299,4/5/2022,Memphis Grizzlies,Utah Jazz,25,28,26,25,26,26.5,26,-105,-130
+300,4/5/2022,Milwaukee Bucks,Chicago Bulls,23,24,22,23,22,24.5,22.8,-125,-105
+301,4/5/2022,New Orleans Pelicans,Sacramento Kings,22,24,22,23,21,21.5,22.4,-135,-105
+302,4/5/2022,Philadelphia 76ers,Indiana Pacers,25,25,25,25,26,24.5,25.2,110,-150
+303,4/5/2022,Portland Trail Blazers,Oklahoma City Thunder,27,30,28,28,25,24.5,27.6,-155,115
+304,4/5/2022,San Antonio Spurs,Denver Nuggets,25,27,25,25,25,24.5,25.4,-105,-135
+305,4/5/2022,Washington Wizards,Minnesota Timberwolves,25,27,26,26,26,24.5,26,-140,105
+306,4/6/2022,Boston Celtics,Chicago Bulls,23,23,23,25,24,24.5,23.6,105,-145
+307,4/6/2022,Brooklyn Nets,New York Knicks,26,25,26,27,25,24.5,25.8,-115,-115
+308,4/6/2022,Dallas Mavericks,Detroit Pistons,26,26,26,27,26,24.5,26.2,-105,-125
+309,4/6/2022,Oklahoma City Thunder,Utah Jazz,27,28,28,31,26,,28,,
+310,4/6/2022,Phoenix Suns,Los Angeles Clippers,25,23,24,26,24,24.5,24.4,-105,-130
+311,4/6/2022,Washington Wizards,Atlanta Hawks,24,24,24,24,25,24.5,24.2,115,-160
+312,4/7/2022,Boston Celtics,Milwaukee Bucks,26,28,27,28,26,27.5,27,-105,-130
+313,4/7/2022,Los Angeles Lakers,Golden State Warriors,25,26,26,26,26,26.5,25.8,-110,-125
+314,4/7/2022,Memphis Grizzlies,Denver Nuggets,23,22,23,24,23,24.5,23,-105,-130
+315,4/7/2022,Orlando Magic,Charlotte Hornets,26,28,29,29,26,26.5,27.6,105,-140
+316,4/7/2022,Philadelphia 76ers,Toronto Raptors,24,25,25,26,25,24.5,25,105,-150
+317,4/7/2022,Portland Trail Blazers,New Orleans Pelicans,22,22,22,24,21,24.5,22.2,110,-150
+318,4/7/2022,San Antonio Spurs,Minnesota Timberwolves,27,26,27,29,26,26.5,27,-105,-130
